+++ src/ui/colors/utils.ts
@@ -1,35 +1,25 @@
+// RGB color in the sRGB color space
 export type RgbColor = {
-  r: number;
-  g: number;
-  b: number;
-  a?: number;
-};
-
-export type OklchColor = {
-  L: number; // 0..1
-  C: number; // 0.. 0.37
-  h: number; //   0..360
+  r: number; // 0..255
+  g: number; // 0..255
+  b: number; // 0..255
   alpha?: number; // 0..1
 };
 
-export type HslColor = {
-  h: number;
-  s: number;
-  l: number;
-  a?: number;
+// Perceptual uniform color, can represent colors outside the sRGB gamut
+export type OklchColor = {
+  L: number; // perceived lightness 0..1
+  C: number; // chroma 0.. 0.37
+  H: number; // hue 0..360
+  alpha?: number; // 0..1
 };
 
-export type LabColor = {
-  L: number;
-  a: number;
-  b: number;
-  alpha?: number;
-};
-
-export type XyzColor = {
-  x: number;
-  y: number;
-  z: number;
+// Perceptual uniform color, can represent colors outside the sRGB gamut
+export type OklabColor = {
+  L: number; // perceived lightness 0..1
+  a: number; // green <-> red -0.4..0.4
+  b: number; // blue <-> yellow -0.4..0.4
+  alpha?: number; // 0..1
 };
 
 export function clampByte(v: number): number {
@@ -72,8 +62,8 @@ export function rgbToHex(_: RgbColor): string {
     .toString(16)
     .slice(1);
 
-  if (_.a !== undefined && _.a < 1.0)
-    hexString += ('00' + Math.round(_.a * 255).toString(16)).slice(-2);
+  if (_.alpha !== undefined && _.alpha < 1.0)
+    hexString += ('00' + Math.round(_.alpha * 255).toString(16)).slice(-2);
 
   // Compress hex from hex-6 or hex-8 to hex-3 or hex-4 if possible
   if (
@@ -86,31 +76,53 @@ export function rgbToHex(_: RgbColor): string {
       hexString[0] +
       hexString[2] +
       hexString[4] +
-      (_.a !== undefined && _.a < 1.0 ? hexString[6] : '');
+      (_.alpha !== undefined && _.alpha < 1.0 ? hexString[6] : '');
   }
 
   return '#' + hexString;
 }
 
-export function luma(color: string): number {
-  const rgb = parseHex(color);
-  if (!rgb) return 0;
-  const [r, g, b] = [rgb.r / 255.0, rgb.g / 255.0, rgb.b / 255.0];
+/**  Generate CSS string for a color */
+export function css(_: string | RgbColor | OklabColor | OklchColor): string {
+  if (typeof _ === 'string') return _;
+  if ('r' in _) return rgbToHex(_);
+  if ('a' in _) {
+    const lab = _ as OklabColor;
+    if ('alpha' in lab && typeof lab.alpha === 'number') {
+      return `lab(${Math.round(lab.L * 1000) / 10}% ${
+        Math.round(lab.a * 100) / 100
+      } ${Math.round(lab.b * 100) / 100} / ${
+        Math.round(lab.alpha * 100) / 100
+      }%)`;
+    }
+    return `lab(${Math.round(lab.L * 1000) / 10}% ${
+      Math.round(lab.a * 100) / 100
+    } ${Math.round(lab.b * 100) / 100})`;
+  }
 
-  // Source: https://www.w3.org/TR/WCAG20/#relativeluminancedef
+  // L is a percentage, C is a number < 0.4, H is a number 0..360
+  // Alpha could be 0..1 or n% (0..100)
+  const oklch = _ as OklchColor;
+  if ('alpha' in oklch && typeof oklch.alpha === 'number') {
+    return `oklch(${Math.round(oklch.L * 1000) / 10}% ${
+      Math.round(oklch.C * 1000) / 1000
+    } ${Math.round(oklch.H * 10) / 10} / ${
+      Math.round(oklch.alpha * 100) / 100
+    }%)`;
+  }
 
-  const conv = (n: number) =>
-    n <= 0.03928 ? n / 12.92 : Math.pow((n + 0.055) / 1.055, 2.4);
-
-  return 0.2126 * conv(r) + 0.7152 * conv(g) + 0.0722 * conv(b);
+  return `oklch(${Math.round(oklch.L * 1000) / 10}% ${
+    Math.round(oklch.C * 1000) / 1000
+  } ${Math.round(oklch.H * 10) / 10})`;
 }
 
+// oklab and oklch:
 // https://bottosson.github.io/posts/oklab/
 
-export function oklchToOklab(_: OklchColor): LabColor {
-  const [L, C, h] = [_.L, _.C, _.h];
-  const hRadians = (h * Math.PI) / 180;
-  const result: LabColor = {
+export function oklchToOklab(_: OklchColor): OklabColor {
+  const [L, C, H] = [_.L, _.C, _.H];
+  const hRadians = (H * Math.PI) / 180;
+  const result: OklabColor = {
     L,
     a: C * Math.cos(hRadians),
     b: C * Math.sin(hRadians),
@@ -119,23 +131,25 @@ export function oklchToOklab(_: OklchColor): LabColor {
   return result;
 }
 
-export function oklabToRgb(_: LabColor): RgbColor {
+export function oklabToOklch(_: OklabColor): OklchColor {
+  const [L, a, b] = [_.L, _.a, _.b];
+  const C = Math.sqrt(a * a + b * b);
+  const hRadians = Math.atan2(b, a);
+  const H = (hRadians * 180) / Math.PI;
+  const result: OklchColor = { L, C, H };
+  if (_.alpha !== undefined) result.alpha = _.alpha;
+  return result;
+}
+
+export function oklabToUnclippedRgb(_: OklabColor): number[] {
   const [l, a, b] = [_.L, _.a, _.b];
-  // const y = (l + 0.16) / 1.16;
-  // const x = a / 1.16 + y;
-  // const z = y - b / 1.16;
-  // const r =
-  //   0.9999999984496163 * x + 0.3963377777753821 * y + 0.2158037572992955 * z;
-  // const g = 0.9999999939972972 * y + -0.105561345615017 * z;
-  // const b2 =
-  //   1.0000000088817607 * x + 2.03211193885992 * y + -0.5226657980972148 * z;
 
   const L = Math.pow(
-    l * 0.9999999984505198 + 0.39633779217376786 * a + 0.2158037580607588 * b,
+    0.9999999984505198 * l + 0.39633779217376786 * a + 0.2158037580607588 * b,
     3
   );
   const M = Math.pow(
-    l * 1.00000000888176 - 0.10556134232365635 * a - 0.0638541747717059 * b,
+    1.00000000888176 * l - 0.10556134232365635 * a - 0.0638541747717059 * b,
     3
   );
   const S = Math.pow(
@@ -150,201 +164,207 @@ export function oklabToRgb(_: LabColor): RgbColor {
   const bl =
     -0.004196086541837188 * L - 0.7034186144594493 * M + 1.7076147009309444 * S;
 
+  // Convert from linear RGB to sRGB
   const conv = (n: number) => {
     const abs = Math.abs(n);
-    if (abs > 0.0031308)
-      return (Math.sign(n) || 1) * (1.055 * Math.pow(abs, 1 / 2.4) - 0.055);
-
-    return n * 12.92;
+    if (abs <= 0.0031308) return n * 12.92;
+    return (Math.sign(n) || 1) * (1.055 * Math.pow(abs, 1 / 2.4) - 0.055);
   };
 
-  return {
-    r: clampByte(conv(r) * 255),
-    g: clampByte(conv(g) * 255),
-    b: clampByte(conv(bl) * 255),
-    a: _.alpha,
-  };
+  return [conv(r), conv(g), conv(bl)];
+}
+
+function inGamut(rgb: number[]): boolean {
+  const [r, g, b] = rgb;
+  return r >= 0 && r <= 1 && g >= 0 && g <= 1 && b >= 0 && b <= 1;
+}
+
+function clampRgb(rgb: number[], alpha?: number): RgbColor {
+  let [r, g, b] = rgb;
+  r = clampByte(r * 255);
+  g = clampByte(g * 255);
+  b = clampByte(b * 255);
+  return alpha !== undefined ? { r, g, b, alpha } : { r, g, b };
+}
+
+/** Convert an oklab color to sRGB, clipping the chroma if necessary */
+export function oklabToRgb(color: OklabColor): RgbColor {
+  let [r, g, b] = oklabToUnclippedRgb(color);
+  if (inGamut([r, g, b])) return clampRgb([r, g, b], color.alpha);
+
+  // Try with chroma = 0
+  const oklch = oklabToOklch(color);
+  oklch.C = 0;
+  [r, g, b] = oklabToUnclippedRgb(oklchToOklab(oklch));
+
+  // If even chroma 0 is not in gamut, return the clamped value
+  if (!inGamut([r, g, b])) return clampRgb([r, g, b], color.alpha);
+
+  // Use a binary search to find a chroma that is in gamut
+  let low = 0;
+  let high = color.L;
+  let mid = (low + high) / 2;
+  oklch.C = mid;
+  const resolution = 0.36 / Math.pow(2, 12);
+  while (high - low > resolution) {
+    mid = (low + high) / 2;
+    oklch.C = mid;
+    [r, g, b] = oklabToUnclippedRgb(oklchToOklab(oklch));
+    if (inGamut([r, g, b])) low = mid;
+    else high = mid;
+  }
+  return clampRgb([r, g, b], color.alpha);
 }
 
 export function oklchToRgb(_: OklchColor): RgbColor {
   return oklabToRgb(oklchToOklab(_));
 }
 
-function hueToRgbChannel(t1: number, t2: number, hue: number): number {
-  if (hue < 0) hue += 6;
-  if (hue >= 6) hue -= 6;
-
-  if (hue < 1) return (t2 - t1) * hue + t1;
-  else if (hue < 3) return t2;
-  else if (hue < 4) return (t2 - t1) * (4 - hue) + t1;
-  return t1;
-}
-
-export function hslToRgb(_: HslColor): RgbColor {
-  let [hue, sat, light] = [_.h, _.s, _.l];
-  hue = ((hue + 360) % 360) / 60.0;
-  light = Math.max(0, Math.min(light, 1.0));
-  sat = Math.max(0, Math.min(sat, 1.0));
-  const t2 = light <= 0.5 ? light * (sat + 1) : light + sat - light * sat;
-  const t1 = light * 2 - t2;
-  return {
-    r: Math.round(255 * hueToRgbChannel(t1, t2, hue + 2)),
-    g: Math.round(255 * hueToRgbChannel(t1, t2, hue)),
-    b: Math.round(255 * hueToRgbChannel(t1, t2, hue - 2)),
-  };
-}
-
-export function rgbToHsl(_: RgbColor): HslColor {
-  const [r, g, b] = [_.r / 255.0, _.g / 255.0, _.b / 255.0];
-  const min = Math.min(r, g, b);
-  const max = Math.max(r, g, b);
-
-  const delta = max - min;
-  let h = +Infinity;
-  let s: number;
-
-  if (max === min) h = 0;
-  else if (r === max) h = (g - b) / delta;
-  else if (g === max) h = 2 + (b - r) / delta;
-  else if (b === max) h = 4 + (r - g) / delta;
-
-  h = Math.min(h * 60, 360);
-
-  if (h < 0) h += 360;
-
-  const l = (min + max) / 2;
-
-  if (max === min) s = 0;
-  else if (l <= 0.5) s = delta / (max + min);
-  else s = delta / (2 - max - min);
-
-  return { h, s, l };
-}
-
-export function saturate(color: string, percent: number): string {
-  const rgb = parseHex(color);
-  if (!rgb) return color;
-  const hsl = rgbToHsl(rgb);
-  hsl.s = Math.min(1, (hsl.s * (100 + percent)) / 100);
-  return rgbToHex(hslToRgb(hsl));
-}
-
-export function lighten(color: string, percent: number): string {
-  const rgb = parseHex(color);
-  if (!rgb) return color;
-  const hsl = rgbToHsl(rgb);
-  hsl.l = Math.min(1, (hsl.l * (100 + percent)) / 100);
-  return rgbToHex(hslToRgb(hsl));
-}
-
-/**
- * L: 0..100
- * a: -128..128
- * b: -128..128
- */
-
-export function labToRgb(_: LabColor): RgbColor {
-  let [L, aStar, bStar] = [_.L, _.a, _.b];
-  L = Math.max(0, Math.min(100, L));
-  aStar = Math.max(-128, Math.min(128, aStar));
-  bStar = Math.max(-128, Math.min(128, bStar));
-  let y = (L + 16) / 116;
-  let x = aStar / 500 + y;
-  let z = y - bStar / 200;
-
-  x = 0.95047 * (x * x * x > 0.008856 ? x * x * x : (x - 16 / 116) / 7.787);
-  y = 1.0 * (y * y * y > 0.008856 ? y * y * y : (y - 16 / 116) / 7.787);
-  z = 1.08883 * (z * z * z > 0.008856 ? z * z * z : (z - 16 / 116) / 7.787);
-
-  let r = x * 3.2406 + y * -1.5372 + z * -0.4986;
-  let g = x * -0.9689 + y * 1.8758 + z * 0.0415;
-  let b = x * 0.0557 + y * -0.204 + z * 1.057;
-
-  r = r > 0.0031308 ? 1.055 * Math.pow(r, 1 / 2.4) - 0.055 : 12.92 * r;
-  g = g > 0.0031308 ? 1.055 * Math.pow(g, 1 / 2.4) - 0.055 : 12.92 * g;
-  b = b > 0.0031308 ? 1.055 * Math.pow(b, 1 / 2.4) - 0.055 : 12.92 * b;
-
-  return {
-    r: clampByte(r * 255),
-    g: clampByte(g * 255),
-    b: clampByte(b * 255),
-  };
-}
-
-/**
- * r: 0..255
- * g: 0..255
- * b: 0..255
- * L: 0..100
- * a: -128..128
- * b: -128..128
- */
-export function rgbToLab(_: RgbColor): LabColor {
-  let [r, g, b] = [_.r, _.g, _.b];
-  r = clampByte(r) / 255;
-  g = clampByte(g) / 255;
-  b = clampByte(b) / 255;
-
-  r = r > 0.04045 ? Math.pow((r + 0.055) / 1.055, 2.4) : r / 12.92;
-  g = g > 0.04045 ? Math.pow((g + 0.055) / 1.055, 2.4) : g / 12.92;
-  b = b > 0.04045 ? Math.pow((b + 0.055) / 1.055, 2.4) : b / 12.92;
-
-  let x = (r * 0.4124 + g * 0.3576 + b * 0.1805) / 0.95047;
-  let y = (r * 0.2126 + g * 0.7152 + b * 0.0722) / 1.0;
-  let z = (r * 0.0193 + g * 0.1192 + b * 0.9505) / 1.08883;
-
-  x = x > 0.008856 ? Math.pow(x, 1 / 3) : 7.787 * x + 16 / 116;
-  y = y > 0.008856 ? Math.pow(y, 1 / 3) : 7.787 * y + 16 / 116;
-  z = z > 0.008856 ? Math.pow(z, 1 / 3) : 7.787 * z + 16 / 116;
-
-  return { L: 116 * y - 16, a: 500 * (x - y), b: 200 * (y - z) };
-}
-
-export function rgbToXyz(_: RgbColor): XyzColor {
+export function rgbToOklab(_: RgbColor): OklabColor {
   const [r, g, b] = [_.r, _.g, _.b];
+
+  // Convert from sRGB to linear RGB
+  const conv = (n: number) => {
+    const abs = Math.abs(n);
+    if (abs <= 0.04045) return n / 12.92;
+    return (Math.sign(n) || 1) * Math.pow((abs + 0.055) / 1.055, 2.4);
+  };
+
+  // const L =
+  //   +0.4121656129659433 * conv(r / 255) +
+  //   0.5362752080314703 * conv(g / 255) +
+  //   0.0514575650074055 * conv(b / 255);
+  // const M =
+  //   +0.21185910790086805 * conv(r / 255) +
+  //   0.6807189586533941 * conv(g / 255) +
+  //   0.1074065798455032 * conv(b / 255);
+  // const S =
+  //   +0.08830979477771493 * conv(r / 255) +
+  //   0.28184741736157754 * conv(g / 255) +
+  //   0.6302613617667958 * conv(b / 255);
+
+  const L =
+    0.41222147079999993 * conv(r / 255) +
+    0.5363325363 * conv(g / 255) +
+    0.0514459929 * conv(b / 255);
+  const M =
+    0.2119034981999999 * conv(r / 255) +
+    0.6806995450999999 * conv(g / 255) +
+    0.1073969566 * conv(b / 255);
+  const S =
+    0.08830246189999998 * conv(r / 255) +
+    0.2817188376 * conv(g / 255) +
+    0.6299787005000002 * conv(b / 255);
+
+  const L3 = Math.cbrt(L);
+  const M3 = Math.cbrt(M);
+  const S3 = Math.cbrt(S);
+
   return {
-    x: 0.430574 * r + 0.34155 * g + 0.178325 * b,
-    y: 0.222015 * r + 0.706655 * g + 0.07133 * b,
-    z: 0.020183 * r + 0.129553 * g + 0.93918 * b,
+    L: 0.2104542553 * L3 + 0.793617785 * M3 - 0.0040720468 * S3,
+    a: 1.9779984951 * L3 - 2.428592205 * M3 + 0.4505937099 * S3,
+    b: 0.0259040371 * L3 + 0.7827717662 * M3 - 0.808675766 * S3,
+    alpha: _.alpha,
   };
 }
 
-export function xyzToRgb(_: XyzColor): { r: number; g: number; b: number } {
-  const [x, y, z] = [_.x, _.y, _.z];
-  return {
-    r: 3.063218 * x - 1.393325 * y - 0.475802 * z,
-    g: -0.969243 * x + 1.875966 * y + 0.041555 * z,
-    b: 0.067871 * x - 0.228834 * y + 1.069251 * z,
-  };
+export function rgbToOklch(_: RgbColor): OklchColor {
+  return oklabToOklch(rgbToOklab(_));
+}
+
+/** Of two foreground colors, return the one with the highest
+ *  contrast ratio with the background
+ */
+export function contrast(
+  background: string,
+  dark?: string,
+  light?: string
+): string {
+  light ??= '#fff';
+  dark ??= '#000';
+
+  const bgColor = parseHex(background)!;
+  const lightContrast = apca(bgColor, parseHex(light)!);
+  const darkContrast = apca(bgColor, parseHex(dark)!);
+
+  return Math.abs(lightContrast) > Math.abs(darkContrast) ? light : dark;
 }
 
 /**
- * Return either dark (default #000) or light (default #fff) depending on
- * the contrast ratio (as per WCAG 2.0 spec)
- * WCAG 2.0
- * - AA
- *     - small text: contrast ratio > 4.5:1
- *     - large text (18px, bold): contrast ratio > 3.1
- * - AAA
- *     - small text: contrast ratio > 7:1
- *     - large text (18px, bold): contrast ratio > 4.5.1
+ * Return a more accurate measure of contrast between a foreground color
+ * and a background color than WCAG2.0.
+ *
+ * If the result is negative, the foreground is lighter than the
+ * background
+ *
+ * Range about -108..108
+ * If abs(result) > 90, suitable for all cases
+ * If abs(result) < 60, large text
+ * If abs(result) < 44, spot and non-text
+ * If abs(result) < 30, minimum contrast for any text
+ *
+ * See https://www.myndex.com/APCA/
  */
-export function contrast(base: string, dark?: string, light?: string): string {
-  let darkContrast, lightContrast;
-  dark ??= '#000';
-  light ??= '#fff';
+export function apca(background: RgbColor, foreground: RgbColor): number {
+  // exponents
+  const normBG = 0.56;
+  const normTXT = 0.57;
+  const revTXT = 0.62;
+  const revBG = 0.65;
 
-  // Calculate contrast ratios for each color
-  // See https://www.w3.org/TR/WCAG20/#contrast-ratiodef
-  const baseLuma = luma(base);
-  const darkLuma = luma(dark);
-  const lightLuma = luma(light);
-  if (baseLuma > darkLuma) darkContrast = (baseLuma + 0.05) / (darkLuma + 0.05);
-  else darkContrast = (darkLuma + 0.05) / (baseLuma + 0.05);
+  // clamps
+  const blkThrs = 0.022;
+  const blkClmp = 1.414;
+  const loClip = 0.1;
+  const deltaYmin = 0.0005;
 
-  if (baseLuma > lightLuma)
-    lightContrast = (baseLuma + 0.05) / (lightLuma + 0.05);
-  else lightContrast = (lightLuma + 0.05) / (baseLuma + 0.05);
+  // scalers
+  // see https://github.com/w3c/silver/issues/645
+  const scaleBoW = 1.14;
+  const loBoWoffset = 0.027;
+  const scaleWoB = 1.14;
+  const loWoBoffset = 0.027;
 
-  return darkContrast > lightContrast ? dark : light;
+  function fclamp(Y: number) {
+    return Y >= blkThrs ? Y : Y + (blkThrs - Y) ** blkClmp;
+  }
+
+  function linearize(val: number) {
+    const sign = val < 0 ? -1 : 1;
+    return sign * Math.pow(Math.abs(val), 2.4);
+  }
+
+  // Calculates "screen luminance" with non-standard simple gamma EOTF
+  // weights should be from CSS Color 4, not the ones here which are via Myndex and copied from Lindbloom
+  const Yfg = fclamp(
+    linearize(foreground.r / 255) * 0.2126729 +
+      linearize(foreground.g / 255) * 0.7151522 +
+      linearize(foreground.b / 255) * 0.072175
+  );
+
+  const Ybg = fclamp(
+    linearize(background.r / 255) * 0.2126729 +
+      linearize(background.g / 255) * 0.7151522 +
+      linearize(background.b / 255) * 0.072175
+  );
+
+  let S: number, C: number, Sapc: number;
+
+  if (Math.abs(Ybg - Yfg) < deltaYmin) C = 0;
+  else {
+    if (Ybg > Yfg) {
+      // dark foreground on light background
+      S = Ybg ** normBG - Yfg ** normTXT;
+      C = S * scaleBoW;
+    } else {
+      // light foreground on dark background
+      S = Ybg ** revBG - Yfg ** revTXT;
+      C = S * scaleWoB;
+    }
+  }
+  if (Math.abs(C) < loClip) Sapc = 0;
+  else if (C > 0) Sapc = C - loWoBoffset;
+  else Sapc = C + loBoWoffset;
+
+  return Sapc * 100;
 }